pyproject.toml
[tool.poetry]
name = "FINESSE"
version = "1.3.0"
description = "A graphical user interface for controlling and monitoring an interferometer device."
authors = [
    "Alex Dewar <a.dewar@imperial.ac.uk>",
    "Diego Alonso Álvarez <d.alonso-alvarez@imperial.ac.uk>",
    "Dan Cummins <daniel.cummins17@imperial.ac.uk>",
    "Callum West <c.west@imperial.ac.uk>",
    "Imperial College London RSE Team <ict-rse-team@imperial.ac.uk>",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
PySide6 = "^6.6.1"
matplotlib = "^3.8.2"
platformdirs = "^4.1.0"
pypubsub = "^4.0.3"
pyyaml = "^6.0"
schema = "^0.7.5"
pyserial = "^3.5"
beautifulsoup4 = "^4.12.2"
python-statemachine = "^2.1.2"
numpy = "^1.26.2"
decorator = "^5.1.1"
pycsvy = "^0.2.2"
frozendict = "^2.3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1.0"
pytest-mypy = "^0.10.1"
pytest-mock = "^3.12.0"
pre-commit = "^3.6.0"
pyinstaller = "^6.3.0"
types-pyyaml = "^6.0.12.12"
types-beautifulsoup4 = "^4.12.0.7"
pytest-qt = "^4.2.0"
types-decorator = "^5.1.8.4"
pydot = "^1.4.2"
ruff = "^0.1.7"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.7.5"
mkdocs-material = "^9.5.2"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"


[tool.ruff]
select = [
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "RUF", # ruff
]
target-version = "py311"

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true

# Allows Type[T] to refer to abstract classes, which is not otherwise supported.
# See https://github.com/python/mypy/issues/4717
disable_error_code = "type-abstract"

[tool.pytest.ini_options]
addopts = "-v --mypy -p no:warnings --cov=finesse --cov-report=html:htmlcov/coverage --doctest-modules --ignore=site/ --ignore=docs/"

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
